move modules into helpers and test-data folders
