+++ src/frontend/efiling-frontend/src/components/page/payment/Payment.stories.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { getDocumentsData } from "../../../modules/documentTestData";
-import { getTestData } from "../../../modules/confirmationPopupTestData";
-import { getCourtData } from "../../../modules/courtTestData";
+import { getDocumentsData } from "../../../modules/test-data/documentTestData";
+import { getTestData } from "../../../modules/test-data/confirmationPopupTestData";
+import { getCourtData } from "../../../modules/test-data/courtTestData";
 
 import Payment from "./Payment";
