+++ src/frontend/efiling-frontend/src/components/page/payment/Payment.test.js
@@ -8,10 +8,10 @@ import {
   getByRole,
   waitFor,
 } from "@testing-library/react";
-import { getTestData } from "../../../modules/confirmationPopupTestData";
-import { getDocumentsData } from "../../../modules/documentTestData";
-import { getCourtData } from "../../../modules/courtTestData";
-import { generateJWTToken } from "../../../modules/authenticationHelper";
+import { getTestData } from "../../../modules/test-data/confirmationPopupTestData";
+import { getDocumentsData } from "../../../modules/test-data/documentTestData";
+import { getCourtData } from "../../../modules/test-data/courtTestData";
+import { generateJWTToken } from "../../../modules/helpers/authenticationHelper";
 
 import Payment from "./Payment";
