+++ src/frontend/efiling-frontend/src/components/page/upload/Upload.test.js
@@ -11,10 +11,10 @@ import {
   queryByText,
   getAllByTestId,
 } from "@testing-library/react";
-import { getTestData } from "../../../modules/confirmationPopupTestData";
-import { getCourtData } from "../../../modules/courtTestData";
-import { getDocumentsData } from "../../../modules/documentTestData";
-import { generateJWTToken } from "../../../modules/authenticationHelper";
+import { getTestData } from "../../../modules/test-data/confirmationPopupTestData";
+import { getCourtData } from "../../../modules/test-data/courtTestData";
+import { getDocumentsData } from "../../../modules/test-data/documentTestData";
+import { generateJWTToken } from "../../../modules/helpers/authenticationHelper";
 
 import Upload, { uploadDocuments } from "./Upload";
