+++ src/frontend/efiling-frontend/src/modules/helpers/authenticationHelper.js
