+++ src/frontend/efiling-frontend/src/modules/helpers/handleBackEvent.js
