+++ src/frontend/efiling-frontend/src/modules/helpers/translateApplicantInfo.js
